Move the search modal to fill the body
