+++ ui/apps/platform/src/Containers/Search/SearchModal.js
@@ -76,7 +76,7 @@ class SearchModal extends Component {
 const SearchModalContainer = (props) => {
     const EnhancedSearchModal = onClickOutside(SearchModal);
     return (
-        <div className="search-modal w-full z-md-300 pf-u-background-color-100">
+        <div className="z-md-300 absolute w-full h-full pf-u-background-color-100">
             <EnhancedSearchModal {...props} />
         </div>
     );
